# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "StellarKraal API",
    "version": "1.0.0",
    "description": "Livestock-backed micro-lending API on Stellar/Soroban. Write endpoints return an unsigned XDR transaction the client must sign with Freighter and submit to the Stellar network.\n\n**Rate limits:** Per-IP limits apply (global, auth, and write tiers). See the [rate limits guide](../docs/guides/rate-limits.md) for limits, headers, and retry behavior. See also the [API quickstart](../docs/guides/api-quickstart.md).\n\n**Changelog:** All API-level additions and breaking changes are documented in the [API Changelog](../docs/api-changelog.md)."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned API"
    },
    {
      "url": "/api",
      "description": "Unversioned (deprecated, redirects to /api/v1)"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/collateral/register": {
      "post": {
        "tags": [
          "collateral"
        ],
        "summary": "Register livestock collateral (on-chain)",
        "description": "Builds an unsigned Soroban transaction to register livestock as on-chain collateral. The client must sign the returned XDR with Freighter and submit it to the Stellar network.",
        "operationId": "registerCollateral",
        "requestBody": {
          "required": true,
          "description": "Livestock collateral details including owner wallet, animal type, count, and appraised value.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCollateralRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unsigned XDR transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XdrResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/collateral/register/batch": {
      "post": {
        "tags": [
          "collateral"
        ],
        "summary": "Batch register livestock collateral (on-chain)",
        "description": "Builds unsigned Soroban transactions for up to 50 livestock collateral registrations in a single request. Returns an array of unsigned XDR transactions in the same order as the input items.",
        "operationId": "batchRegisterCollateral",
        "requestBody": {
          "required": true,
          "description": "An array of 1–50 collateral registration items to process in batch.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRegisterCollateralRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Array of unsigned XDR transactions, one per item in input order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRegisterCollateralResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/collateral": {
      "post": {
        "tags": [
          "collateral"
        ],
        "summary": "Create collateral record",
        "description": "Creates a livestock collateral record from JSON data or a multipart registration with animal details and a JPEG or PNG image. This does not submit an on-chain transaction; use POST /collateral/register for on-chain registration.",
        "operationId": "createCollateral",
        "requestBody": {
          "required": true,
          "description": "Collateral record details including animal type, count, and appraised value.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1CreateCollateralRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/V1MultipartCollateralRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollateralRecord"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "get": {
        "tags": [
          "collateral"
        ],
        "summary": "List collateral records",
        "description": "Returns a paginated list of collateral records. Supports filtering by owner wallet address and animal type. Results are ordered by creation date descending.",
        "operationId": "listCollateral",
        "security": [],
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "description": "Filter by Stellar public key of the collateral owner.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "animal_type",
            "in": "query",
            "description": "Filter by animal type (e.g. \"cattle\", \"goat\", \"sheep\").",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (1-indexed).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number of records per page (1–100).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollateralPage"
                }
              }
            }
          }
        }
      }
    },
    "/collateral/{id}": {
      "get": {
        "tags": [
          "collateral"
        ],
        "summary": "Get collateral by ID",
        "description": "Returns a single collateral record by its unique identifier. Returns 404 if the record does not exist.",
        "operationId": "getCollateral",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique collateral record identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collateral record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollateralRecord"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "collateral"
        ],
        "summary": "Partially update collateral fields",
        "description": "Updates one or more fields on an existing collateral record. At least one of animal_type, count, or appraised_value must be provided. Returns the updated record.",
        "operationId": "updateCollateral",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique collateral record identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Fields to update. At least one field must be provided.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1UpdateCollateralRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollateralRecord"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "delete": {
        "tags": [
          "collateral"
        ],
        "summary": "Soft-delete a collateral record",
        "description": "Marks a collateral record as deleted without removing it from the database. The record can be restored via PATCH /collateral/{id}/restore. Returns the deletion confirmation.",
        "operationId": "deleteCollateral",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique collateral record identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/loan/request": {
      "post": {
        "tags": [
          "loan"
        ],
        "summary": "Request a new loan",
        "description": "Builds an unsigned Soroban transaction to request a new loan against registered collateral. Consider supplying an `Idempotency-Key` header on retries to avoid accidentally submitting duplicate loan requests.",
        "operationId": "requestLoan",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional client-generated UUID v4 key. If supplied, the server stores the response for 24 hours and returns the original on replay with `X-Idempotent-Replayed: true`."
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Loan request payload including borrower public key, collateral ID, and requested amount.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unsigned XDR transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XdrResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/loan/liquidate": {
      "post": {
        "tags": [
          "loan"
        ],
        "summary": "Liquidate an undercollateralised loan",
        "description": "Builds an unsigned Soroban transaction to liquidate a loan whose health factor has fallen below the liquidation threshold. The liquidator repays part of the loan and receives the pledged collateral. Returns 400 with the current health factor if the loan is not yet liquidatable.",
        "operationId": "liquidateLoan",
        "requestBody": {
          "required": true,
          "description": "Liquidation details including liquidator wallet, loan ID, and repayment amount.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanLiquidateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unsigned XDR transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XdrResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/webhooks/{id}": {
      "delete": {
        "tags": [
          "webhooks"
        ],
        "summary": "Delete a webhook",
        "description": "Permanently removes a registered webhook by its ID. Returns 204 with no body on success, or 404 if the webhook does not exist.",
        "operationId": "deleteWebhook",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique webhook identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Webhook deleted (no body)"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 JWT issued by POST /api/auth/login. Required on all POST, PUT, DELETE, and PATCH requests except /api/auth/* endpoints. Pass as: Authorization: Bearer <token>"
      }
    },
    "schemas": {
...
      "XdrResponse": {
        "type": "object",
        "description": "An unsigned Soroban transaction encoded as Base64 XDR. The client must sign this with Freighter and submit to the Stellar network.",
        "properties": {
          "api_version": {
            "type": "string",
            "description": "API version identifier.",
            "example": "v1"
          },
          "xdr": {
            "type": "string",
            "description": "Base64-encoded unsigned Soroban transaction XDR"
          }
        }
      },
...
    },
...
  }
}